feat(commonware-node): catch execution layer up to consensus on start (#307)
Forwards already-finalized blocks from consensus to execution layer if
holes are detected on startup.

Building on #281 and #305, this patch solves the problem of the
execution layer missing blocks due to reth not flushing some blocks to
disk on SIGINT/SIGTERM.

On startup, if `latest_consensus_finalized_height >
latest_execution_block_number`, then the execution driver spawns a task
to read all already finalized blocks for heights `h` such that
`latest_consensus_finalized_height <= h < latest_execution_bock_number`
from the consensus marshal agent and to make them available to the
execution layer. Blocks are read in order of descending height to allow
the node to participate in consensus verification/proposal as soon as
possible.

In line with #305, the latest finalized block is only kept for actual
new finalization at the tip of the chain. Backfilled blocks do not
update the value.